refactor: Refatoração do componente Overlay
